feat: modularization
modularization of the project and application of the console library was carried out.
